Some bug fixes.
Fixed a couple of bug related to TreeItem where i forgot to link it with the instance of the Tag.
Plus remove the possibility to set the communication type for a tag but only if is local (Which need to still be implemented) or not. If not, will allow to change the address.