More Tag UI & Comm String Address Improvements.
Improved the UI for the tags when renaming the tag, now the left padding for tree table view is respected. Added possibility to add tag/folder from a folder and clone a communication tag via a context menu.
Remove the "readOnly" properties for string address data since now, with tags, is not required.
Now the tag stage can be open from the MainEditStage.