Fixed a couple of bugs. Improvements and cleanup.
Fixed a bug where CommThread would accept an update if the PLC is not connected stay in update forever.
Done some refractoring (Move TagStage inside HMIManager).
Implemented correctly the "isActive" method inside the TagData for the AttributeProperty.
Also change the key for the newValueRunnable inside the ValueIntermediate to be an object, so you can use anything as a key.